Module Easy_format.Compact

No spacing or newlines other than those in the input data or those produced by Custom printing.

val to_buffer : Buffer.t -> t -> unit

val to_string : t -> string

val to_channel : out_channel -> t -> unit

val to_stdout : t -> unit

val to_stderr : t -> unit

val to_formatter : Format.formatter -> t -> unit